requirements.txt

streamlit
pandas
requests
beautifulsoup4
tldextract
phonenumbers
dnspython
googlesearch-python
tqdm
